added slash comamnd support

+++ src/main/java/de/kittybot/kittybot/modules/PaginatorModule.java
@@ -2,11 +2,13 @@ package de.kittybot.kittybot.modules;
 
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
+import de.kittybot.kittybot.command.context.CommandContext;
 import de.kittybot.kittybot.module.Module;
 import de.kittybot.kittybot.objects.Emoji;
 import de.kittybot.kittybot.objects.Paginator;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.Permission;
+import net.dv8tion.jda.api.entities.MessageEmbed;
 import net.dv8tion.jda.api.entities.TextChannel;
 import net.dv8tion.jda.api.events.message.guild.react.GuildMessageReactionAddEvent;
 import org.jetbrains.annotations.NotNull;
@@ -53,9 +55,11 @@ public class PaginatorModule extends Module{
 			}
 		}
 		else if(Emoji.WASTEBASKET.get().equals(code)){
-			event.getChannel().deleteMessageById(event.getMessageIdLong()).queue();
-			this.paginators.invalidate(event.getMessageIdLong());
-			return;
+			if(paginator.getAuthorId() == event.getUserIdLong()){
+				event.getChannel().deleteMessageById(event.getMessageIdLong()).queue();
+				this.paginators.invalidate(event.getMessageIdLong());
+				return;
+			}
 		}
 		event.getReaction().removeReaction(event.getUser()).queue();
 	}
@@ -72,13 +76,26 @@ public class PaginatorModule extends Module{
 		channel.clearReactionsById(paginator.getMessageId()).queue();
 	}
 
-	public void create(TextChannel channel, int maxPages, BiFunction<Integer, EmbedBuilder, EmbedBuilder> embedFunction){
-		var embedBuilder = embedFunction.apply(0, new EmbedBuilder().setFooter("Page 1/" + maxPages)).build();
+	public void create(TextChannel channel, long authorId, int maxPages, BiFunction<Integer, EmbedBuilder, EmbedBuilder> embedFunction){
+		var embedBuilder = embedFunction.apply(0, new EmbedBuilder().setFooter("Page: 1/" + maxPages)).build();
+		create(maxPages, embedFunction, embedBuilder, channel, authorId);
+	}
+
+	public void create(CommandContext ctx, int maxPages, BiFunction<Integer, EmbedBuilder, EmbedBuilder> embedFunction){
+		ctx.acknowledge(true);
+		var embedBuilder = embedFunction.apply(0, new EmbedBuilder().setFooter("Page: 1/" + maxPages)).build();
+		var channel = ctx.getChannel();
+		create(maxPages, embedFunction, embedBuilder, channel, ctx.getUserId());
+	}
+
+	public void create(int maxPages, BiFunction<Integer, EmbedBuilder, EmbedBuilder> embedFunction, MessageEmbed embedBuilder, TextChannel channel, long userId){
 		channel.sendMessage(embedBuilder).queue(message -> {
-			var paginator = new Paginator(message, maxPages, embedFunction);
-			if(maxPages > 0 && channel.getGuild().getSelfMember().hasPermission(Permission.MESSAGE_HISTORY, Permission.MESSAGE_ADD_REACTION)){
-				message.addReaction(Emoji.ARROW_LEFT.get()).queue();
-				message.addReaction(Emoji.ARROW_RIGHT.get()).queue();
+			var paginator = new Paginator(message, userId, maxPages, embedFunction);
+			if(channel.getGuild().getSelfMember().hasPermission(Permission.MESSAGE_HISTORY, Permission.MESSAGE_ADD_REACTION)){
+				if(maxPages > 1){
+					message.addReaction(Emoji.ARROW_LEFT.get()).queue();
+					message.addReaction(Emoji.ARROW_RIGHT.get()).queue();
+				}
 				message.addReaction(Emoji.WASTEBASKET.get()).queue();
 			}
 			this.paginators.put(paginator.getMessageId(), paginator);
